refactor: move v2 components to v1 folder
