+++ src/pages/checkout/[taskId].tsx
@@ -1,7 +1,7 @@
 import { useRouter } from "next/router";
 
 import { CheckoutPage } from "@Billing/Checkout";
-import { FeatureFlag } from "@Components/utils/FeatureFlag";
+import { FeatureFlag } from "src/components/utils/FeatureFlag";
 
 const CheckoutRoute = () => {
   const router = useRouter();
